refactor(tests): consolidate Worktree Module — 13 files → 2
Closes #3742
